-Just some testing. -Added a few things. Performence helpers. -Added better Sprinkler Particles. -Still tring to fix Sync Issus. Darn MC is fucking hard with that. -Adjusted the Pipe Redstone pipes for small flowing states (10mB per tick is the limit). Will be changed soon to another redstone pipe. -Added the Mod: AdvGenerators. You do not see any source heh? Yeah no open source support. Sorry. Thats it.

+++ Spmod_Common/Source/speiger/src/api/common/world/tiles/interfaces/IHopper.java
@@ -45,7 +45,7 @@ public abstract interface IHopper
 	/**
 	 * @return the HopperFacing
 	 */
-	public abstract int getFacing();
+	public abstract short getFacing();
 	
 	/**
 	 * @return if the Hopper has a Rotation
@@ -55,7 +55,7 @@ public abstract interface IHopper
 	/**
 	 * @return the Hopper Rotation
 	 */
-	public abstract int getRotation();
+	public abstract short getRotation();
 	
 	/**
 	 * @return if the hopper has a inventory
