+++ Spmod_Common/Source/speiger/src/tinymodularthings/common/blocks/storage/TinyChest.java
@@ -166,7 +166,7 @@ public class TinyChest extends TileFacing implements IInventory,
 	@Override
 	public void onPlaced(int facing)
 	{
-		setFacing(facing);
+		setFacing((short)facing);
 		if (isValidMode())
 		{
 			updateInventory();
